Crate hello_rust

This is where everything learning rust programming begins.

This is a rust project called hello_rust that is self-documenting to teach rust programming. You are now reading an example of rust crate comment that is placed at the top of a file called main.rs in the root of the directory of this project.

You can view the source for this file by clicking on source on the upper right corner. In the source code notice the start /*! and end markers */. These determine which comments are crate or module level comments.

You can see below that each module that is part of this crate is listed under the Modules header. You can access each of these chapters by clicking model names,such as chapter_00.

Within each module/crate document there will also be additional information about the functions within that module. This crate has a single function called main and you can see the document about it below under the Functions header.

After reading this entire page and the main function details you can continue your rust learning journey by clicking on the first module chapter_00

Modules

chapter_00

This chapter is only about one thing; getting code running in rust.

chapter_01

Variables and basic data types in rust

chapter_02

Functions

main

function fn main() is the entry point to the binary executable. ie your end goal of a running program. fn main(){…} is where your program starts running. All comments starting with /** on a line of its own for public functions will be part of your documentation. Now click on the name of the function main to see more details about the main function.