Module hello_rust::chapter_01

The following is how you use another 'crate', a crate is the name for a rust executable or library. 'std' is the identifier for the standard rust crate that comes with rust by default. Any rust project can use the std crate without the need for the import statement which otherwise would be needed to import a foreign crate :

extern crate std

In the std crate there quite a few parts called modules. One of them is the std::env module that helps you access the environment of your program. To use env we need the following line:

use std::env;
mod point; 

Besides the foreign crates, your project code can have different modules. In the above line you can see how to access the point module from the main.rs file that we are in. It enables access to the public contents of the point.rs file which is in the same directory.

As you can see rust comes up a full development environment, even the documentation support comes out of the box. And your code documentation becomes an easy to read standard HTML format that you can share as you please.

Not only that, but you can even write unit test statements into code comments, which the cargo test command will execute.

Functions

test_struct
Let's create a struct

testfn
will this be part of documents ?