Function hello_rust::chapter_01::run

pub fn run()

Each chapter will have a run function that will be called from main.rs.

These function will execute sample code relevant to each chapter.

Arguments

None

Remarks

This function will introduce some variables. Using let x = 2 is a simple way to create a variable on stack. Rust infers the local variables types, but you can also specify.

All variables are immutable by default. To have a variable that you can change you have to use mut keyword.