Function hello_rust::main

pub fn main()

function fn main() is the entry point to the binary executable. ie your end goal of a running program. fn main(){…} is where your program starts running. All comments starting with /** on a line of its own for public functions will be part of your documentation. Now click on the name of the function main to see more details about the main function.

Arguments

None

You can write details about a functions arguments details using # Arguments

Remarks

You can write remarks for your functions using # Remarks

Examples

You can also write code examples to share using # Examples

println!("Hello World");  

In the main function we will call another function from a module named chapter_00. To use anything from another module we need to import that module first. We import the module chapter_00 on line number 15 of the main.rs file.

pub mod chapter_00;